--- a/Fable.Remoting.AspNetCore/Middleware.fs
+++ b/Fable.Remoting.AspNetCore/Middleware.fs
@@ -137,7 +137,7 @@ module internal Middleware =
                   ctx.Response.StatusCode <- 200
                   ctx.Response.ContentType <- "application/octet-stream"
                   use ms = new MemoryStream ()
-                  Fable.Remoting.MsgPack.Write.object output ms
+                  Fable.Remoting.MsgPack.Write.serializeObj output ms
                   ms.Position <- 0L
                   do! ms.CopyToAsync ctx.Response.Body
                   return! next ctx