--- a/Fable.Remoting.MsgPack/paket.references
+++ b/Fable.Remoting.MsgPack/paket.references
@@ -1,1 +1,2 @@
 FSharp.Core
+TypeShape