--- a/paket.dependencies
+++ b/paket.dependencies
@@ -11,6 +11,7 @@ nuget NUnit
 nuget PuppeteerSharp
 nuget Suave
 nuget Giraffe
+nuget TypeShape
 
 group Build
     source https://api.nuget.org/v3/index.json