@Zaid-Ajaj 2020-07-25
> I don't really know what's going on, but this incantation works and it's faaaaaaaaaaast

I don't either but I like what I am seeing (the insane numbers), I have seen TypeShape before but I haven't used it myself. Though the `TypeInfo` API in SimpleJson is somewhat inspired by how TypeShape looks like. Let's get this in and optimize Fable.Remoting.Server later :smile: