@kerams 2020-07-26
I really appreciate the gesture, but I am content with the status quo. However, if you need a helping hand maintaining this repo, which would be understandable given how many projects you have and how pristine you seem to keep the issue and PR lists, then I'd be willing to step up. I'm on Fable's gitter, you could drop me a private message there.