@Zaid-Ajaj 2020-07-26
>  However, if you need a helping hand maintaining this repo

That was my intention with sending the invitation, having another maintainer would really really help me out, especially someone who is familiar with the codebase :wink: